move hash_payload to access ALGORITHMS
